# apps/desktop/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZETA_VN Desktop Integration API",
    "version": "1.0.0",
    "description": "API tích hợp Desktop (Electron/React) ↔ AI Server"
  },
  "paths": {
...
    "/api/v1/training/jobs": {
      "post": {
        "summary": "Create Training Job",
        "operationId": "create_training_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJob"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Training Jobs",
        "operationId": "list_training_jobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingJob"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training/jobs/{job_id}": {
      "get": {
        "summary": "Get Training Job",
        "operationId": "get_training_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJob"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TrainingJobCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "data_source": {
            "type": "string"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["name"]
      },
      "TrainingJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["created", "running", "paused", "completed", "cancelled", "failed"]
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "data_source": {
            "type": "string"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          },
          "logs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["id", "name", "status", "progress", "created_at", "updated_at", "config", "logs"]
      },
...
    }
  }
}