# apps/desktop/src/api/generated/models/TrainingJobCreate.ts
import Record from "Record";
import TrainingJobCreate from "./TrainingJobCreate";
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TrainingJobCreate = {
    name: string;
    data_source?: string;
    config?: Record<string, any>;
};